requirements.txt
python-dotenv==0.21.1
pytz==2022.7.1
Telethon==1.26.1